Clean imports, move tests to same package
